Update wordpress-multi-server.yaml
Changed the label for wp_master_server_flavor from Server Size to Master Server Size and wp_web_server_flavor to Node Server Size because it was confusing in Reach. Both showed Server Size as labels and it was hard to tell which server you were changing the size of.