Adding load balancer config.
This is not in a working state right now, things are in flux.
To get secondary instance IPs we must use a child template; only
top-level attributes are available for in-line resource group created
resources, so it's not possible to get the IPs for secondaries with this
single-template setup.

We will also need to update pillar data on the secondaries and
re-converge salt at add the memcache php session module config.